Provide generic version of StringUtils::Split accepting an output iterator
The previous version always split strings to a vector, which may
or may not be the best container depending on the kind of data
that is split. For flags for example, a std::set seems more natural.
The generic versions SplitTo() accept the same arguments as Split()
but additionally expect an output iterator to put elements into.
Split() was also moved to use SplitTo() to avoid code duplication.