Do not display windowed mode setting on incompatible platforms
The user was previously able to decide whether the "Windowed" mode
option was shown in the settings with the XML "canwindowed" setting.
This would show the option even when the windowing system was not
actually capable of producing windowed output, leading to undefined
behavior and a deadlock.

This change makes the canwindowed a strict disable-only setting, i.e.
the user can disable being offered windowed modes, but cannot enable
them when there is no actual support implemented.

The #ifdef in AdvancedSettings for iOS is not needed anymore, since
CanDoWindowed() will already return false for the iOS windowing
implementation.