Add parameter to GraphicContext::SetVideoResolution to control whether to call into windowing
When windowing implementations have to update the video resoultion
due to outside events (e.g. monitor hot-plug), they currently have
to call SetVideoResolution. But there is no reason why GraphicContext
should then call back into the windowing system again. This introduces
a new parameter that controls whether SetVideoResolution should call
windowing or not. It is expected to be set to true when Kodi updates
the resolution e.g. for video playback and set to false when windowing
updates the resolution, but that is not a requirement. Windowing
implementations can use the parameter as they see fit.

Signed-off-by: Philipp Kerling <pkerling@casix.org>